Fair comparison between Ls.
Before, I had the different CSMs reconstructed using different Ls
running on the same timeseries, but different substrings of it (the
longer the L, the shorter the substring). This gave the larger L CSMs
an artificial advantage, since they were predicting on shorter
substrings and thus had *less* chances to make mistakes. Now, all of
the reconstructed CSMs predict on the *same* length during tuning.
Thus, we're making sure we compare apples to apples with the tuning
performance.