Jerry-rigged filtering code.
After realizing that CSSR doesn't *all* possible histories that might
belong to a state, I jerry-rigged the code so that it

(1) synchronizes by using a lookup of histories and then

(2) continues ahead with the filtering by using the topology of the
reconstructed epsilon machine, until a disallowed transition occurs, at
which point we go back to (1).

This isn't the best possible solution, but it will have to do for now.

The completed fix is in filter_data_fixing1.py, which, on all future
commits (not including this one), will go back to being called
filter_data.py.