Incorporate Zero-order CSM.
This is a baseline test. Basically, we predict as if the process is a
biased coin. So if the person mostly tweets, we always tweet, and vice
versa. This is like using the sample mean of an output as the predictor
(since, barring any other information, this *is* the optimal predictor).