Change simulation and filtering code.
I changed the simulation code to output a data file as well as a raster
plot for a given causal state model.

I changed the filtering code so that I can specify whether or not I
want to compare the predicted and true timeseries, to avoid cluttering
up the Terminal window.