Modified filtering code, both the methods and the single files.
In the methods code, I added an option to print (or not) the filtered
state series to the command line.

In the single file, I modified the code so that instead of providing
the user rank, one provides the user_id that you would like to run. I
also changed the test lines to include / not include the predictions
and filtered states.